refactor(sensitive-storage): separate vulnerable and secure contracts into distinct modules
Split sensitive_storage contract into canonical pattern:
- Create src/secure.rs with SecureStorageContract (hash-only storage)
- Remove inline secure methods from lib.rs
- Add #[contracttype] DataKey enum to secure module
- Implement re-init guard and admin auth validation
- Add 3 comprehensive tests for secure variant

Secure improvements:
- initialize() stores only commitment hash, never raw secret
- get_commitment() returns hash without auth requirement
- No get_secret() method in secure variant
- Re-init guard prevents admin replacement

Update documentation:
- Clarify reference path to src/secure.rs
- Add 'Why it is dangerous' section
- Explain off-chain key management pattern

Test results: All 5 tests pass (2 vulnerable + 3 secure)
Build: Clean (no errors)
Format: Passes cargo fmt check
