+++ vulnerable/sensitive_storage/src/lib.rs
@@ -7,17 +7,18 @@
 //! VULNERABILITY: `initialize()` writes raw secret material to persistent
 //! storage, exposing it to anyone who can read the ledger.
 //!
-//! SECURE MIRROR: store only a SHA-256 hash commitment on-chain; keep the
-//! raw secret in off-chain key management infrastructure.
+//! SECURE MIRROR: see `secure.rs` — stores only a SHA-256 hash commitment
+//! on-chain; raw secret never touched by the contract.
 
 #![no_std]
 use soroban_sdk::{contract, contractimpl, contracttype, Address, Bytes, Env};
+
+mod secure;
 
 #[contracttype]
 pub enum DataKey {
     Admin,
     SecretKey,
-    Commitment,
 }
 
 #[contract]
@@ -53,28 +54,6 @@ impl SensitiveStorageContract {
             .persistent()
             .get(&DataKey::Admin)
             .expect("admin not initialized")
-    }
-
-    // -------------------------------------------------------------------------
-    // Secure mirror — stores only a hash commitment, never the raw secret.
-    // -------------------------------------------------------------------------
-
-    /// SECURE: stores only a hash commitment — the raw secret never touches the ledger.
-    pub fn initialize_secure(env: Env, admin: Address, secret_hash: Bytes) {
-        admin.require_auth();
-        env.storage().persistent().set(&DataKey::Admin, &admin);
-        // ✅ Only the hash is stored — the raw secret never touches the ledger
-        env.storage()
-            .persistent()
-            .set(&DataKey::Commitment, &secret_hash);
-    }
-
-    /// Returns the stored hash commitment.
-    pub fn get_commitment(env: Env) -> Bytes {
-        env.storage()
-            .persistent()
-            .get(&DataKey::Commitment)
-            .expect("commitment not set")
     }
 }
 
@@ -116,27 +95,5 @@ mod tests {
         // No auth mock needed — get_secret has no access control
         let stolen = client.get_secret();
         assert_eq!(stolen, secret);
-    }
-
-    #[test]
-    fn test_secure_stores_only_commitment() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let contract_id = env.register_contract(None, SensitiveStorageContract);
-        let client = SensitiveStorageContractClient::new(&env, &contract_id);
-
-        let admin = Address::generate(&env);
-        // Simulate a SHA-256 hash (32 bytes) — raw secret never sent on-chain
-        let hash = Bytes::from_slice(&env, &[0xab_u8; 32]);
-
-        client.initialize_secure(&admin, &hash);
-
-        let stored_commitment = client.get_commitment();
-        assert_eq!(stored_commitment, hash);
-        // Raw secret is never stored — DataKey::SecretKey is absent
-        let has_secret_key = env.as_contract(&contract_id, || {
-            env.storage().persistent().has(&DataKey::SecretKey)
-        });
-        assert!(!has_secret_key);
     }
 }
